api/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
backports.entry-points-selectable==1.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
codegen==1.0
colorama==0.4.6
coverage==7.2.7
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.12.2
idna==3.4
jedi==0.18.2
mailjet-rest==1.3.4
null==0.6.1
packaging==23.1
parso==0.8.3
patterns==0.3
Pillow==10.0.0
pip-review==1.3.0
pipenv==2023.7.3
plan==0.5
platformdirs==3.8.0
prompt-toolkit==3.0.38
ptpython==3.0.23
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
